starting the process of using conn and fill_surface through out filters
