+++ src/porespy/filters/_funcs.py
@@ -8,11 +8,19 @@ import scipy.ndimage as spim
 from skimage.morphology import reconstruction
 from skimage.segmentation import clear_border
 from skimage.morphology import ball, disk, square, cube, diamond, octahedron
-from porespy.tools import _check_for_singleton_axes
-from porespy.tools import get_border, subdivide, recombine
-from porespy.tools import unpad, extract_subsection
-from porespy.tools import ps_disk, ps_ball, ps_round
-from porespy.tools import get_tqdm, get_edt
+from porespy.tools import (
+    _check_for_singleton_axes,
+    get_border,
+    subdivide,
+    recombine,
+    unpad,
+    extract_subsection,
+    ps_disk,
+    ps_ball,
+    ps_round,
+    get_tqdm,
+    get_edt,
+)
 from porespy import settings
 from typing import Literal
 
@@ -23,12 +31,7 @@ __all__ = [
     "apply_padded",
     "chunked_func",
     "distance_transform_lin",
-    "fill_blind_pores",
-    "find_disconnected_voxels",
     "find_dt_artifacts",
-    "find_surface_pores",
-    "find_hidden_pores",
-    "find_invalid_pores",
     "flood",
     "flood_func",
     "hold_peaks",
@@ -37,11 +40,6 @@ __all__ = [
     "porosimetry",
     "prune_branches",
     "region_size",
-    "trim_disconnected_blobs",
-    "trim_extrema",
-    "trim_floating_solid",
-    "trim_nonpercolating_paths",
-    "trim_small_clusters",
 ]
 
 
@@ -51,7 +49,13 @@ logger = logging.getLogger(__name__)
 strel = {2: {'min': disk(1), 'max': square(3)}, 3: {'min': ball(1), 'max': cube(3)}}
 
 
-def apply_padded(im, pad_width, func, pad_val=1, **kwargs):
+def apply_padded(
+    im,
+    pad_width,
+    func,
+    pad_val=1,
+    **kwargs,
+):
     r"""
     Applies padding to an image before sending to ``func``, then extracts
     the result corresponding to the original image shape.
@@ -61,8 +65,8 @@ def apply_padded(im, pad_width, func, pad_val=1, **kwargs):
     im : ndarray
         The image to which ``func`` should be applied
     pad_width : int or list of ints
-        The amount of padding to apply to each axis. Refer to
-        ``numpy.pad`` documentation for more details.
+        The amount of padding to apply to each axis. Refer to ``numpy.pad``
+        documentation for more details.
     pad_val : scalar
         The value to place into the padded voxels.  The default is 1 (or
         ``True``) which extends the pore space.
@@ -90,42 +94,11 @@ def apply_padded(im, pad_width, func, pad_val=1, **kwargs):
     return result
 
 
-def trim_small_clusters(im, size=1):
-    r"""
-    Remove isolated voxels or clusters of a given size or smaller
-
-    Parameters
-    ----------
-    im : ndarray
-        The binary image from which voxels are to be removed.
-    size : scalar
-        The threshold size of clusters to trim.  As clusters with this
-        many voxels or fewer will be trimmed.  The default is 1 so only
-        single voxels are removed.
-
-    Returns
-    -------
-    im : ndarray
-        A copy of ``im`` with clusters of voxels smaller than the given
-        ``size`` removed.
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/filters/reference/trim_small_clusters.html>`_
-    to view online example.
-
-    """
-    strel = ps_round(r=1, ndim=im.ndim, smooth=False)
-    filtered_array = np.copy(im)
-    labels, N = spim.label(filtered_array, structure=strel)
-    id_sizes = np.array(spim.sum(im, labels, range(N + 1)))
-    area_mask = id_sizes <= size
-    filtered_array[area_mask[labels]] = 0
-    return filtered_array
-
-
-def hold_peaks(im, axis=-1, ascending=True):
+def hold_peaks(
+    im,
+    axis=-1,
+    ascending=True,
+):
     r"""
     Replaces each voxel with the highest value along the given axis.
 
@@ -198,16 +171,17 @@ def distance_transform_lin(
     mode : str
         Controls how the distance is measured. Options are:
 
-        'forward'
-            Distances are measured in the increasing direction
-            along the specified axis
-        'reverse'
-            Distances are measured in the reverse direction.
-            'backward' is also accepted.
-        'both'
-            Distances are calculated in both directions (by
-            recursively calling itself), then reporting the minimum value
-            of the two results.
+        ========== =================================================================
+        Mode       Description
+        ========== =================================================================
+        'forward'  Distances are measured in the increasing direction along the
+                   specified axis
+        'reverse'  Distances are measured in the reverse direction. 'backward' is
+                   also accepted.
+        'both'     Distances are calculated in both directions (by recursively
+                   calling itself), then reporting the minimum value of the two
+                   results.
+        ========== =================================================================
 
     Returns
     -------
@@ -252,306 +226,11 @@ def distance_transform_lin(
     return f
 
 
-def find_disconnected_voxels(im, conn: int = None, surface: bool = False):
-    r"""
-    Identifies all voxels that are not connected to the edge of the image.
-
-    Parameters
-    ----------
-    im : ndarray
-        A Boolean image, with ``True`` values indicating the phase for which
-        disconnected voxels are sought.
-    conn : int
-        For 2D the options are 4 and 8 for square and diagonal neighbors,
-        while for the 3D the options are 6 and 26, similarily for square
-        and diagonal neighbors. The default is the maximum option.
-    surface : bool
-        If ``True`` any isolated regions touching the edge of the image are
-        considered disconnected.
-
-    Returns
-    -------
-    image : ndarray
-        An ndarray the same size as ``im``, with ``True`` values indicating
-        voxels of the phase of interest (i.e. ``True`` values in the original
-        image) that are not connected to the outer edges.
-
-    See Also
-    --------
-    fill_blind_pores, trim_floating_solid
-
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/filters/reference/find_disconnected_voxels.html>`_
-    to view online example.
-
-    """
-    _check_for_singleton_axes(im)
-
-    if im.ndim == 2:
-        if conn == 4:
-            strel = disk(1)
-        elif conn in [None, 8]:
-            strel = square(3)
-        else:
-            raise Exception("Received conn is not valid")
-    elif im.ndim == 3:
-        if conn == 6:
-            strel = ball(1)
-        elif conn in [None, 26]:
-            strel = cube(3)
-        else:
-            raise Exception("Received conn is not valid")
-    labels, N = spim.label(input=im, structure=strel)
-    if not surface:
-        holes = clear_border(labels=labels) > 0
-    else:
-        keep = set(np.unique(labels))
-        for ax in range(labels.ndim):
-            labels = np.swapaxes(labels, 0, ax)
-            keep.intersection_update(set(np.unique(labels[0, ...])))
-            keep.intersection_update(set(np.unique(labels[-1, ...])))
-            labels = np.swapaxes(labels, 0, ax)
-        holes = np.isin(labels, list(keep), invert=True)
-    return holes
-
-
-def find_hidden_pores(im, conn='min'):
-    r"""
-    Finds hidden pores that a not connected to any surface
-
-    Parameters
-    ----------
-    im : ndarray
-        A boolean array with `True` indicating the phase of interest
-    conn : str
-        Can be either `min` or `max` and controls the shape of the structuring
-        element used to determine voxel connectivity.  The default if `'min'` which
-        imposes the strictest criteria, so that voxels must share a face to be
-        considered connected.
-
-    Returns
-    -------
-    hidden : ndarray
-        A array containing boolean values indicating voxels which belong to hidden
-        pores.
-    """
-    from porespy.generators import borders
-    se = strel[im.ndim][conn]
-    labels, N = spim.label(input=im, structure=se)
-    mask = borders(im.shape, mode='faces')
-    hits = np.unique(labels[mask])
-    hidden = np.isin(labels, hits, invert=True)
-    return hidden
-
-
-def find_surface_pores(im, conn='min'):
-    r"""
-    Finds surface pores that do not span the domain
-
-    Parameters
-    ----------
-    im : ndarray
-        A boolean array with `True` indicating the phase of interest
-    conn : str
-        Can be either `min` or `max` and controls the shape of the structuring
-        element used to determine voxel connectivity.  The default if `'min'` which
-        imposes the strictest criteria, so that voxels must share a face to be
-        considered connected.
-
-    Returns
-    -------
-    surface : ndarray
-        A array containing boolean values indicating voxels which belong to surface
-        pores.
-    """
-    se = strel[im.ndim][conn]
-    labels, N = spim.label(input=im, structure=se)
-    keep = set()
-    for ax in range(labels.ndim):
-        labels = np.swapaxes(labels, 0, ax)
-        s1 = set(np.unique(labels[0, ...]))
-        s2 = set(np.unique(labels[-1, ...]))
-        tmp = s1.intersection(s2)
-        keep.update(tmp)
-        labels = np.swapaxes(labels, 0, ax)
-    hidden = find_hidden_pores(im, conn=conn)
-    surface = np.isin(labels, list(keep), invert=True) * ~hidden
-    return surface
-
-
-def find_invalid_pores(im, conn='min'):
-    r"""
-    Finds invalid pores which are either hidden or do not span the domain
-
-    Parameters
-    ----------
-    im : ndarray
-        A boolean array with `True` indicating the phase of interest
-    conn : str
-        Can be either `min` or `max` and controls the shape of the structuring
-        element used to determine voxel connectivity.  The default if `'min'` which
-        imposes the strictest criteria, so that voxels must share a face to be
-        considered connected.
-
-    Returns
-    -------
-    invalid : ndarray
-        A array containing `1` indicated hidden pores and `2` indicating surface
-        pores.
-    """
-    hidden = find_hidden_pores(im=im, conn=conn)
-    surface = find_surface_pores(im=im, conn=conn)
-    invalid = hidden.astype(int) + 2*surface.astype(int)
-    return invalid
-
-
-def fill_blind_pores(im, conn: int = None, surface: bool = False):
-    r"""
-    Fills all blind pores that are isolated from the main void space.
-
-    Parameters
-    ----------
-    im : ndarray
-        The image of the porous material
-
-    Returns
-    -------
-    im : ndarray
-        A Boolean image, with `True` values indicating the phase of interest.
-    conn : int
-        For 2D the options are 4 and 8 for square and diagonal neighbors,
-        while for the 3D the options are 6 and 26, similarily for square
-        and diagonal neighbors. The default is the maximum option.
-    surface : bool
-        If `True`, any isolated pore regions that are connected to the
-        sufaces of the image are but not connected to the main percolating
-        path are also removed. When this is enabled, only the voxels
-        belonging to the largest region are kept. This can be
-        problematic if image contains non-intersecting tube-like structures,
-        for instance, since only the largest tube will be preserved.
-
-    Returns
-    -------
-    im : ndarray
-        A version of `im` but with all the blind or disconnected pores converted
-        to solid (i.e. `False`)
-
-    See Also
-    --------
-    find_disconnected_voxels
-    trim_nonpercolating_paths
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/filters/reference/fill_blind_pores.html>`_
-    to view online example.
-
-    """
-    im = np.copy(im)
-    holes = find_disconnected_voxels(im, conn=conn, surface=surface)
-    im[holes] = False
-    return im
-
-
-def trim_floating_solid(im, conn: int = None, surface: bool = False):
-    r"""
-    Removes all solid that that is not attached to main solid structure.
-
-    Parameters
-    ----------
-    im : ndarray
-        The image of the porous material
-    conn : int
-        For 2D the options are 4 and 8 for square and diagonal neighbors,
-        while for the 3D the options are 6 and 26, similarily for square
-        and diagonal neighbors. The default is the maximum option.
-    surface : bool
-        If ``True``, any isolated solid regions that are connected to the
-        surfaces of the image but not the main body of the solid are also
-        removed.  When this is enabled, only the voxels belonging to the
-        largest region are kept. This can be problematic if the image
-        contains non-intersecting tube-like structures, for instance,
-        since only the largest tube will be preserved.
-
-    Returns
-    -------
-    image : ndarray
-        A version of ``im`` but with all the disconnected solid removed.
-
-    See Also
-    --------
-    find_disconnected_voxels
-    trim_nonpercolating_paths
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/filters/reference/trim_floating_solid.html>`_
-    to view online example.
-
-    """
-    im = np.copy(im)
-    holes = find_disconnected_voxels(~im, conn=conn, surface=surface)
-    im[holes] = True
-    return im
-
-
-def trim_nonpercolating_paths(im, inlets, outlets, strel=None):
-    r"""
-    Remove all nonpercolating paths between specified locations
-
-    Parameters
-    ----------
-    im : ndarray
-        The image of the porous material with ```True`` values indicating the
-        phase of interest
-    inlets : ndarray
-        A boolean mask indicating locations of inlets, such as produced by
-        ``porespy.generators.faces``.
-    outlets : ndarray
-        A boolean mask indicating locations of outlets, such as produced by
-        ``porespy.generators.faces``.
-    strel : ndarray
-        The structuring element to use when determining if regions are
-        connected.  This is passed to ``scipiy.ndimage.label``.
-
-    Returns
-    -------
-    image : ndarray
-        A copy of ``im`` with all the nonpercolating paths removed
-
-    Notes
-    -----
-    This function is essential when performing transport simulations on an
-    image since regions that do not span between the desired inlet and
-    outlet do not contribute to the transport.
-
-    See Also
-    --------
-    find_disconnected_voxels
-    trim_floating_solid
-    trim_blind_pores
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/filters/reference/trim_nonpercolating_paths.html>`_
-    to view online example.
-
-    """
-    labels = spim.label(im, structure=strel)[0]
-    IN = np.unique(labels * inlets)
-    OUT = np.unique(labels * outlets)
-    hits = np.array(list(set(IN).intersection(set(OUT))))
-    new_im = np.isin(labels, hits[hits > 0])
-    return new_im
-
-
-def trim_extrema(im, h, mode="maxima"):
+def trim_extrema(
+    im,
+    h,
+    mode="maxima",
+):
     r"""
     Trims local extrema in greyscale values by a specified amount.
 
@@ -602,7 +281,7 @@ def flood(
     im,
     labels,
     mode: Literal['maximum', 'minimum', 'median', 'mean', 'size',
-                  'standard_deviations', 'variance'] = "max",
+                  'standard_deviations',  'variance'] = "max",
 ):
     r"""
     Floods/fills each region in an image with a single value based on the
@@ -670,7 +349,11 @@ def flood(
     return flooded
 
 
-def flood_func(im, func, labels=None):
+def flood_func(
+    im,
+    func,
+    labels=None,
+):
     r"""
     Flood each isolated region in an image with a constant value calculated by
     the given function.
@@ -682,19 +365,19 @@ def flood_func(im, func, labels=None):
         and 0's elsewhere.
     func : Numpy function handle
         The function to be applied to each region in the image.  Any Numpy
-        function that returns a scalar value can be passed, such as ``amin``,
-        ``amax``, ``sum``, ``mean``, ``median``, etc.
+        function that returns a scalar value can be passed, such as `amin`,
+        `amax`, `sum`, `mean`, `median`, etc.
     labels : ndarray
         An array containing labels identifying each individual region to be
-        flooded. If not provided then ``scipy.ndimage.label`` is applied to
-        ``im > 0``.
+        flooded. If not provided then `scipy.ndimage.label` is applied to
+        `im > 0`.
 
     Returns
     -------
     flooded : ndarray
-        An image the same size as ``im`` with each isolated region flooded
-        with a constant value based on the given ``func`` and the values
-        in ``im``.
+        An image the same size as `im` with each isolated region flooded
+        with a constant value based on the given `func` and the values
+        in `im`.
 
     See Also
     --------
@@ -702,8 +385,8 @@ def flood_func(im, func, labels=None):
 
     Notes
     -----
-    Many of the functions in ``scipy.ndimage`` can be applied to
-    individual regions using the ``index`` argument.  This function extends
+    Many of the functions in `scipy.ndimage` can be applied to
+    individual regions using the `index` argument.  This function extends
     that behavior to all numpy function, in the event you wanted to compute
     the cosine of the values in each region for some reason. This function
     also floods the original image instead of returning a list of values for
@@ -740,7 +423,7 @@ def find_dt_artifacts(dt):
     Returns
     -------
     image : ndarray
-        An ndarray the same shape as ``dt`` with numerical values
+        An ndarray the same shape as `dt` with numerical values
         indicating the maximum amount of error in each volxel, which is
         found by subtracting the distance to nearest edge of image from
         the distance transform value. In other words, this is the error
@@ -770,27 +453,32 @@ def find_dt_artifacts(dt):
     return result
 
 
-def region_size(im, strel=None):
+def region_size(
+    im,
+    conn='min',
+):
     r"""
     Replace each voxel with the size of the region to which it belongs
 
     Parameters
     ----------
     im : ndarray
-        Either a boolean image wtih ``True`` indicating the features of
-        interest, in which case ``scipy.ndimage.label`` will be applied to
+        Either a boolean image wtih `True` indicating the features of
+        interest, in which case `scipy.ndimage.label` will be applied to
         find regions, or a greyscale image with integer values indicating
         regions.
-    strel : ndarray
-        The structuring element to use for defining which connected voxels form
-        a region
+    conn : str
+        Can be either `'min'` or `'max'` and controls the shape of the structuring
+        element used to determine voxel connectivity.  The default if `'min'` which
+        imposes the strictest criteria, so that voxels must share a face to be
+        considered connected.
 
     Returns
     -------
     image : ndarray
-        A copy of ``im`` with each voxel value indicating the size of the
+        A copy of `im` with each voxel value indicating the size of the
         region to which it belongs.  This is particularly useful for
-        finding chord sizes on the image produced by ``apply_chords``.
+        finding chord sizes on the image produced by `apply_chords`.
 
     See Also
     --------
@@ -798,7 +486,7 @@ def region_size(im, strel=None):
 
     Notes
     -----
-    This function provides the same result as ``flood`` with ``mode='size'``,
+    This function provides the same result as `flood` with `mode='size'`,
     although does the computation in a different way.
 
     Examples
@@ -808,8 +496,9 @@ def region_size(im, strel=None):
     to view online example.
 
     """
+    se = strel[im.ndim][conn]
     if im.dtype == bool:
-        im = spim.label(im, structure=strel)[0]
+        im = spim.label(im, structure=se)[0]
     counts = np.bincount(im.flatten())
     counts[0] = 0
     return counts[im]
@@ -828,27 +517,27 @@ def apply_chords(
     Parameters
     ----------
     im : ndarray
-        An image of the porous material with void marked as ``True``.
+        An image of the porous material with void marked as `True`.
     spacing : int
         Separation between chords.  The default is 1 voxel.  This can be
         decreased to 0, meaning that the chords all touch each other,
-        which automatically sets to the ``label`` argument to ``True``.
+        which automatically sets to the `label` argument to `True`.
     axis : int (default = 0)
         The axis along which the chords are drawn.
-    trim_edges : bool (default = ``True``)
+    trim_edges : bool (default = `True`)
         Whether or not to remove chords that touch the edges of the image.
         These chords are artifically shortened, so skew the chord length
         distribution.
-    label : bool (default is ``False``)
-        If ``True`` the chords in the returned image are each given a
+    label : bool (default is `False`)
+        If `True` the chords in the returned image are each given a
         unique label, such that all voxels lying on the same chord have
-        the same value.  This is automatically set to ``True`` if spacing
-        is 0, but is ``False`` otherwise.
+        the same value.  This is automatically set to `True` if spacing
+        is 0, but is `False` otherwise.
 
     Returns
     -------
     image : ndarray
-        A copy of ``im`` with non-zero values indicating the chords.
+        A copy of `im` with non-zero values indicating the chords.
 
     See Also
     --------
@@ -884,7 +573,11 @@ def apply_chords(
     return result
 
 
-def apply_chords_3D(im, spacing: int = 0, trim_edges: bool = True):
+def apply_chords_3D(
+    im,
+    spacing: int = 0,
+    trim_edges: bool = True,
+):
     r"""
     Adds chords to the void space in all three principle directions.
 
@@ -893,11 +586,11 @@ def apply_chords_3D(im, spacing: int = 0, trim_edges: bool = True):
     Parameters
     ----------
     im : ndarray
-        A 3D image of the porous material with void space marked as True.
+        A 3D image of the porous material with void space marked as `True`.
     spacing : int (default = 0)
         Chords are automatically separed by 1 voxel on all sides, and this
         argument increases the separation.
-    trim_edges : bool (default is ``True``)
+    trim_edges : bool (default is `True`)
         Whether or not to remove chords that touch the edges of the image.
         These chords are artifically shortened, so skew the chord length
         distribution
@@ -905,7 +598,7 @@ def apply_chords_3D(im, spacing: int = 0, trim_edges: bool = True):
     Returns
     -------
     image : ndarray
-        A copy of ``im`` with values of 1 indicating x-direction chords,
+        A copy of `im` with values of 1 indicating x-direction chords,
         2 indicating y-direction chords, and 3 indicating z-direction
         chords.
 
@@ -913,7 +606,7 @@ def apply_chords_3D(im, spacing: int = 0, trim_edges: bool = True):
     -----
     The chords are separated by a spacing of at least 1 voxel so that
     tools that search for connected components, such as
-    ``scipy.ndimage.label`` can detect individual chords.
+    `scipy.ndimage.label` can detect individual chords.
 
     See Also
     --------
@@ -967,22 +660,23 @@ def local_thickness(
     mode : str
         Controls with method is used to compute the result. Options are:
 
-        'hybrid'
-            (default) Performs a distance transform of the void
-            space, thresholds to find voxels larger than ``sizes[i]``, trims
-            the resulting mask if ``access_limitations`` is ``True``, then
-            dilates it using the efficient fft-method to obtain the
-            non-wetting fluid configuration.
-        'dt'
-            Same as 'hybrid', except uses a second distance transform,
-            relative to the thresholded mask, to find the invading fluid
-            configuration. The choice of 'dt' or 'hybrid' depends on speed,
-            which is system and installation specific.
-        'mio'
-            Using a single morphological image opening step to obtain
-            the invading fluid confirguration directly, *then* trims if
-            ``access_limitations`` is ``True``. This method is not ideal and
-            is included for comparison purposes.
+        ============ ===============================================================
+        Mode         Description
+        ============ ===============================================================
+        'hybrid'     (default) Performs a distance transform of the void space,
+                     thresholds to find voxels larger than `sizes[i]`, trims
+                     the resulting mask if ``access_limitations`` is ``True``, then
+                     dilates it using the efficient fft-method to obtain the
+                     non-wetting fluid configuration.
+        'dt'         Same as 'hybrid', except uses a second distance transform,
+                     relative to the thresholded mask, to find the invading fluid
+                     configuration. The choice of 'dt' or 'hybrid' depends on speed,
+                     which is system and installation specific.
+        'mio'        Using a single morphological image opening step to obtain
+                     the invading fluid confirguration directly, *then* trims if
+                     ``access_limitations`` is ``True``. This method is not ideal
+                     and is included for comparison purposes.
+        ============ ===============================================================
 
     divs : int or array_like
         The number of times to divide the image for parallel processing.  If ``1``
@@ -993,7 +687,7 @@ def local_thickness(
     Returns
     -------
     image : ndarray
-        A copy of ``im`` with the pore size values in each voxel.
+        A copy of `im` with the pore size values in each voxel.
 
     See Also
     --------
@@ -1064,23 +758,24 @@ def porosimetry(
     mode : str
         Controls with method is used to compute the result. Options are:
 
-        'hybrid'
-            (default) Performs a distance tranform of the void
-            space, thresholds to find voxels larger than ``sizes[i]``,
-            trims the resulting mask if ``access_limitations`` is ``True``,
-            then dilates it using the efficient fft-method to obtain the
-            non-wetting fluid configuration.
-        'dt'
-            Same as 'hybrid', except uses a second distance
-            transform, relative to the thresholded mask, to find the
-            invading fluid configuration. The choice of 'dt' or 'hybrid'
-            depends on speed, which is system and installation specific.
-        'mio'
-            Uses binary erosion followed by dilation to obtain the invading
-            fluid configuration directly. If ``access_limitated`` is
-            ``True`` then disconnected blobs are trimmmed before the dilation.
-            This is the only method that can be parallelized by chunking (see
-            ``divs`` and ``cores``).
+        ============ ===============================================================
+        Mode         Description
+        ============ ===============================================================
+        'hybrid'     (default) Performs a distance tranform of the void
+                     space, thresholds to find voxels larger than ``sizes[i]``,
+                     trims the resulting mask if ``access_limitations`` is ``True``,
+                     then dilates it using the efficient fft-method to obtain the
+                     non-wetting fluid configuration.
+        'dt'         Same as 'hybrid', except uses a second distance
+                     transform, relative to the thresholded mask, to find the
+                     invading fluid configuration. The choice of 'dt' or 'hybrid'
+                     depends on speed, which is system and installation specific.
+        'mio'        Uses binary erosion followed by dilation to obtain the invading
+                     fluid configuration directly. If ``access_limitated`` is
+                     ``True`` then disconnected blobs are trimmmed before the
+                     dilation. This is the only method that can be parallelized by
+                     chunking (see ``divs`` and ``cores``).
+        ============ ===============================================================
 
     divs : int or array_like
         The number of times to divide the image for parallel processing.
@@ -1210,84 +905,19 @@ def porosimetry(
     return imresults
 
 
-def trim_disconnected_blobs(im, inlets, strel=None):
-    r"""
-    Removes foreground voxels not connected to specified inlets.
-
-    Parameters
-    ----------
-    im : ndarray
-        The image containing the blobs to be trimmed
-    inlets : ndarray or tuple of indices
-        The locations of the inlets.  Can either be a boolean mask the
-        same shape as ``im``, or a tuple of indices such as that returned
-        by the ``where`` function.  Any voxels *not* connected directly to
-        the inlets will be trimmed.
-    strel : array-like
-        The neighborhood over which connectivity should be checked. It
-        must be symmetric and the same dimensionality as the image. It is
-        passed directly to the ``scipy.ndimage.label`` function as the
-        ``structure`` argument so refer to that docstring for additional
-        info.
-
-    Returns
-    -------
-    image : ndarray
-        An array of the same shape as ``im``, but with all foreground
-        voxels not connected to the ``inlets`` removed.
-
-    See Also
-    --------
-    find_disconnected_voxels, find_nonpercolating_paths
-
-    Examples
-    --------
-    `Click here
-    <https://porespy.org/examples/filters/reference/trim_disconnected_blobs.html>`_
-    to view online example.
-
-    """
-    if isinstance(inlets, tuple):
-        temp = np.copy(inlets)
-        inlets = np.zeros_like(im, dtype=bool)
-        inlets[temp] = True
-    elif (inlets.shape == im.shape) and (inlets.max() == 1):
-        inlets = inlets.astype(bool)
-    else:
-        raise Exception("inlets not valid, refer to docstring for info")
-    if strel is None:
-        if im.ndim == 3:
-            strel = cube(3)
-        else:
-            strel = square(3)
-    labels = spim.label(inlets + (im > 0), structure=strel)[0]
-    keep = np.unique(labels[inlets])
-    keep = keep[keep > 0]
-    im2 = np.isin(labels, keep)
-    im2 = im2 * im
-    return im2
-
-
-def _get_axial_shifts(ndim=2, include_diagonals=False):
+def _get_axial_shifts(ndim=2, conn='min'):
     r"""
     Helper function to generate the axial shifts that will be performed on
     the image to identify bordering pixels/voxels
     """
+    neighbors = strel[ndim][conn]
     if ndim == 2:
-        if include_diagonals:
-            neighbors = square(3)
-        else:
-            neighbors = diamond(1)
         neighbors[1, 1] = 0
         x, y = np.where(neighbors)
         x -= 1
         y -= 1
         return np.vstack((x, y)).T
     else:
-        if include_diagonals:
-            neighbors = cube(3)
-        else:
-            neighbors = octahedron(1)
         neighbors[1, 1, 1] = 0
         x, y, z = np.where(neighbors)
         x -= 1
@@ -1296,7 +926,7 @@ def _get_axial_shifts(ndim=2, include_diagonals=False):
         return np.vstack((x, y, z)).T
 
 
-def _make_stack(im, include_diagonals=False):
+def _make_stack(im, conn='min'):
     r"""
     Creates a stack of images with one extra dimension to the input image
     with length equal to the number of borders to search + 1.
@@ -1307,7 +937,7 @@ def _make_stack(im, include_diagonals=False):
 
     """
     ndim = len(np.shape(im))
-    axial_shift = _get_axial_shifts(ndim, include_diagonals)
+    axial_shift = _get_axial_shifts(ndim, conn)
     if ndim == 2:
         stack = np.zeros([np.shape(im)[0], np.shape(im)[1], len(axial_shift) + 1])
         stack[:, :, 0] = im
@@ -1328,7 +958,7 @@ def _make_stack(im, include_diagonals=False):
         return stack
 
 
-def nphase_border(im, include_diagonals=False):
+def nphase_border(im, conn='min'):
     r"""
     Identifies the voxels in regions that border *N* other regions.
 
@@ -1340,9 +970,11 @@ def nphase_border(im, include_diagonals=False):
         An ND image of the porous material containing discrete values in
         the pore space identifying different regions. e.g. the result of a
         snow-partition
-    include_diagonals : bool
-        When identifying bordering pixels (2D) and voxels (3D) include
-        those shifted along more than one axis
+    conn : str
+        Can be either `'min'` or `'max'` and controls the shape of the structuring
+        element used to determine voxel connectivity.  The default if `'min'` which
+        imposes the strictest criteria, so that voxels must share a face to be
+        considered connected.
 
     Returns
     -------
@@ -1365,7 +997,7 @@ def nphase_border(im, include_diagonals=False):
     # Pad image to handle edges
     im = np.pad(im, pad_width=1, mode="edge")
     # Stack rolled images for each neighbor to be inspected
-    stack = _make_stack(im, include_diagonals)
+    stack = _make_stack(im, conn)
     # Sort the stack along the last axis
     stack.sort()
     out = np.ones_like(im)
@@ -1414,10 +1046,7 @@ def prune_branches(skel, branch_points=None, iterations: int = 1):
 
     """
     skel = skel > 0
-    if skel.ndim == 2:
-        from skimage.morphology import square as cube
-    else:
-        from skimage.morphology import cube
+    cube = strel[skel.ndim]['max']
     # Create empty image to house results
     im_result = np.zeros_like(skel)
     # If branch points are not supplied, attempt to find them
